fel: Fix USB timeout on large transfers
Trying to use oversized initrd files (20 MB or more) can fail
with the "libusb usb_bulk_send error -1" error message.

To address this problem, we can split the transfer into smaller
chunks and the problem disappears. Effectively, this is a revert
of the older "fel: Increase timeout to 60 seconds instead of
splitting bulk transfers" commmit.

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>