fel: Make the SPL load address configurable instead of hardcoded 0x0
Add a new field 'spl_addr' to the SoC description structure and adjust
the code to honor it. This is needed for supporting Allwinner A80.

Tested on Allwinner A20 by changing the 'spl_addr' to 0x28000 in the
'fel' tool and using a custom build of U-Boot (CONFIG_SPL_TEXT_BASE
changed from 0x20 to 0x28020).

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>